fix: 同步导出下载方式调整 # Reviewed, transaction id: 3061
